+++ index.html
@@ -20,10 +20,6 @@
       content="initial-scale=1.0, user-scalable=no, maximum-scale=1, width=device-width"
     />
     <link rel="manifest" href="/manifest.json" />
-    <script
-      type="text/javascript"
-      src="//dapi.kakao.com/v2/maps/sdk.js?appkey=<%= kakaoAppKey %>&libraries=services"
-    ></script>
     <title>Milideal</title>
   </head>
   <body>
