+++ src/pages/Map/hooks/index.ts
@@ -1,0 +1,4 @@
+import useMap from "./useMap";
+import useMarkers from "./useMarkers";
+
+export { useMap, useMarkers };
